% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateandtime.r
\name{.ad}
\alias{.ad}
\title{shortcut}
\usage{
.ad(x, ...)
}
\arguments{
\item{x}{date input}

\item{...}{other input sent to as.Date()}
}
\description{
shortcut for as.Date...
}
\examples{
.ad(.date(),format="\%Y\%m\%d")   
}
