% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.adf}
\alias{.adf}
\title{.adf}
\usage{
.adf(matrix, colnames = NULL, rownames = NULL)
}
\arguments{
\item{matrix}{input matrix}

\item{colnames}{colnames}

\item{rownames}{rownames}
}
\description{
as.data.frame
}
