% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.adfm}
\alias{.adfm}
\title{shortcut}
\usage{
.adfm(...)
}
\arguments{
\item{...}{R object}
}
\value{
a data frame.
}
\description{
convert a vector to a data.frame (of possibly different columns).
}
\examples{
.adfm(1:10,ncol=2)
}
