% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.alrna}
\alias{.alrna}
\title{shortcut}
\usage{
.alrna(integer, names = NULL)
}
\arguments{
\item{integer}{integer indicating the length of the list}

\item{names}{name of each element of the list (default is NULL)}
}
\value{
an object of class 'list'
}
\description{
creates a list of (possibly named) elements
}
\examples{
.alrna(10,LETTERS[1:10])
}
