% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.an}
\alias{.an}
\title{shortcut}
\usage{
.an(..., warning = FALSE)
}
\arguments{
\item{...}{R object}

\item{warning}{suppress warnings}
}
\value{
a vector of numerical values.
}
\description{
shortcut of as.numeric().
}
\examples{
.an("2")
}
