% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.annt}
\alias{.annt}
\title{shortcut}
\usage{
.annt(vector)
}
\arguments{
\item{vector}{vector object}
}
\value{
a vector of numerical values.
}
\description{
as.numeric of the names of table applied to a vector
}
\examples{
.an("2")    
}
