% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.r
\name{.ar}
\alias{.ar}
\title{id files based arrays}
\usage{
.ar(id1, id2, id3 = NULL, id4 = NULL, id5 = NULL, value = NA)
}
\arguments{
\item{id1}{first id file}

\item{id2}{second id file}

\item{id3}{third id file}

\item{id4}{fourth id file}

\item{id5}{fifth id file}

\item{value}{content of the array, NA by default}
}
\value{
an array (or a matrix when only 2 id files are provided)
}
\description{
creates an array with dim(ensions) and dimnames coming from id files, typically generated via .idf().
}
\examples{
    
\dontrun{       
.idf(table(rpois(10,3)),"X")
.idf(LETTERS[1:6],"Y")   
.ar(id.X,id.Y)    
}   
}
