% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{.axis}
\alias{.axis}
\title{.axis}
\usage{
.axis(
  data,
  split,
  side = 3,
  at.y = NULL,
  col.abline = "gray",
  col.label = "blue",
  col.axis = "blue",
  padj = NA,
  hadj = NA,
  cex = 1,
  add.abline = TRUE,
  add.label = TRUE,
  add.axis = FALSE,
  label = NULL,
  ...
)
}
\arguments{
\item{data}{a file ('id.t'-like] with a column 'pos' (position on the xaxis),
typically 1 to nrow for a split at the full file level}

\item{split}{a factor (typically year/month, school/class, aso)}

\item{side}{3 = over, 1 = below}

\item{at.y}{if non-null, label are displayed in the plot on an horizontal line
at 'at.y' on the y-axis}

\item{col.abline}{colour of the vertical boundaries in abline() \link{col}}

\item{col.label}{colour of the label in axis() \link{col.axis}}

\item{col.axis}{colour of the x-axis if add.axis == TRUE in axis \link{col.axis}}

\item{padj}{padj}

\item{hadj}{hadj}

\item{cex}{cex}

\item{add.abline}{TRUE for vertical boundaries}

\item{add.label}{TRUE for label}

\item{add.axis}{TRUE for x-axis}

\item{label}{label}

\item{...}{...}
}
\description{
Helper for axis
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{OTHER useful arguments [sometimes passed through '...']
cex: cex of labels in axis() and text() [cex.axis]
lty: lty in abline()
padj (hadj): adjustment of the location of the label in axis()
orientation of the label in axis()
orientation of the label in text() if at.y != NULL
y-position of the x-axis in axis()
}\if{html}{\out{</div>}}
}
