% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installpackages.r
\name{.bc}
\alias{.bc}
\title{shorcut}
\usage{
.bc(..., ask = TRUE)
}
\arguments{
\item{...}{string indicating the name of the package to install}

\item{ask}{logical indicating whether to prompt user before installed packages are updated (argument of BiocManager::install)}

\item{dependencies}{logical for dependencies}

\item{quiet}{logical for messages}
}
\description{
function allowing to install and load bioconductor packages
}
