% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.cat}
\alias{.cat}
\title{.cat}
\usage{
.cat(i, n.loop, n.dotline = 50, msg = NULL)
}
\arguments{
\item{i}{iteration number}

\item{n.loop}{number of loop}

\item{n.dotline}{number dot per line}

\item{msg}{message text}
}
\description{
cat a dot or iteration number in a loop
}
