% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapulse.r
\name{.cbf}
\alias{.cbf}
\title{.cbf}
\usage{
.cbf(list.files, path = F, name.loaded.file = "mx.g.zt")
}
\arguments{
\item{list.files}{list of files to bind}

\item{path}{path=T: on doit loader un fichier; path=F: le fichier est dans .GlobalEnv}

\item{name.loaded.file}{name.loaded.file = name of the loaded file}
}
\description{
cbind.files.fun used in radio and tv-research
}
