% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{.circle}
\alias{.circle}
\title{.circle}
\usage{
.circle(x, plot = TRUE, ...)
}
\arguments{
\item{x}{x is a vector of:
x\link{1}: x-coordinate
x\link{2}: y-coordinate
x\link{3}: radius}

\item{plot}{plot}

\item{...}{...}
}
\description{
.circle
}
