% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{.col}
\alias{.col}
\title{colour shortcut function}
\usage{
.col()
}
\value{
a function generating the colours of interest
}
\description{
create a function generating the colours or interest
}
\examples{
\donttest{
pal.fun = .col()
}
}
