% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{.color}
\alias{.color}
\title{colour shortcut function}
\usage{
.color()
}
\value{
a function generating the colours of interest
}
\description{
create a function generating the colours or interest
}
\examples{
\dontrun{
pal.fun = .color()
}
}
