% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.r
\name{.coord}
\alias{.coord}
\title{.coord}
\usage{
.coord(n.dk, n.tk)
}
\arguments{
\item{n.dk}{avec d pour desire et k = 1...n.k}

\item{n.tk}{avec t pour total  et k = 1...n.k}
}
\value{
sum(temp)
}
\description{
.coord
}
