% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.r
\name{.coordr}
\alias{.coordr}
\title{.coordr}
\usage{
.coordr(kd, n.tk)
}
\arguments{
\item{kd}{position dans k dim desiree   (1 < kd < prod(n.tk))}

\item{n.tk}{vecteur des dim de l-array de base (n.tk avec t pour total  et k = 1...n.k )}
}
\value{
out
}
\description{
.coordr
}
