% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateandtime.r
\name{.date}
\alias{.date}
\title{iso date function}
\usage{
.date(x)
}
\description{
indicate the date of the day (the function is used) in iso format (typically used to name output files).
}
\examples{
.date()    
}
