% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{.ellipse}
\alias{.ellipse}
\title{bivariate Gaussian ellipse function}
\usage{
.ellipse(mu = c(0, 0), sigma = diag(2), quantile = 0.95, plot = TRUE, ...)
}
\arguments{
\item{mu}{vector of length 2 indicating the coordinates of the centre of the ellipse. (0,0) by default}

\item{sigma}{2-by-2 covariance matrix (with identity as default)}

\item{quantile}{quantile of the bivariate Gaussian to display}

\item{plot}{a logical. use TRUE to plot or FALSE to save dataset of x and y coordinates of 200 points}

\item{...}{...}
}
\description{
plot an ellipse based on coordinates of centre, covariance matrix and quantile of the bivariate gaussian. Note that visualisation may be affected by the aspect ratio of the plot (argument 'asp' of plot)
}
\examples{
\dontrun{
.ep(asp=1)
data = mvtnorm::rmvnorm(500,c(-1,2),sigma=matrix(c(1,.5,.5,2),ncol=2))
plot(data[,1],data[,2])
.ellipse(apply(data,2,mean),var(data),col="red",lty=2)
.ellipse(c(-1,2),matrix(c(1,.5,.5,2),ncol=2),col="blue",lty=1)
}
}
