% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{.ep}
\alias{.ep}
\title{.ep}
\usage{
.ep(xlim = c(0, 1), ylim = c(0, 1))
}
\arguments{
\item{xlim}{xlim}

\item{ylim}{ylim}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{First define a function generating the colors you want
pal.fun = .col() # pick what is suitable (n doesn't matter)
then use it.
}\if{html}{\out{</div>}}
}
