% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.eval}
\alias{.eval}
\title{.eval}
\usage{
.eval(expr, ...)
}
\arguments{
\item{expr}{expression}

\item{...}{other parameters}
}
\description{
evaluate an R expression based on an input of class 'character'.
Returns the output of the evaluation, whatever it is.
}
\examples{
.eval('c(1,2)')
}
