% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.expit}
\alias{.expit}
\title{shortcut}
\usage{
.expit(x)
}
\arguments{
\item{x}{vector of rational numbers}
}
\value{
a vector of rational numbers belonging to (0,1)
}
\description{
defines the expit function on a vector of numerical values. typically used to define the probabilities of success related to a logistic regression.
}
\examples{
.expit(0)
}
