% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.fill}
\alias{.fill}
\title{.fill
for title, add stuff}
\usage{
.fill(x, npos = NULL, with = " ", front = TRUE)
}
\arguments{
\item{x}{input R object}

\item{npos}{npos}

\item{with}{with}

\item{front}{front}
}
\description{
.fill
for title, add stuff
}
