% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.h}
\alias{.h}
\title{print function for arrays}
\usage{
.h(x, size = c(10, 10, 2))
}
\arguments{
\item{x}{R object to}

\item{size}{size of the output}
}
\description{
prints the first dimensions of an array (up to size 3), matrix or data frame. Less busy people may consider the function head().
}
\examples{
dimw = c(10,20,30) 
arw  = array(1:prod(dimw),dimw)
.h(arw) 
.h(.adf(arw[,,1]))
}
