% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{.hist}
\alias{.hist}
\title{Non-paramteric density plots}
\usage{
.hist(
  data,
  h = 1,
  col = NULL,
  border = NULL,
  lwd = 1,
  points = TRUE,
  first = "points"
)
}
\arguments{
\item{data}{A vector or matrix of numerical values.}

\item{h}{A numeric (data is a vector) or vector of numerical values (data is a matrix) corresponding to the smoothing parameter to use. Default to 1. When \code{h=NULL}, the smoothing value is optimised in the \link[sm]{sm.density} function (Good luck with that).}

\item{col}{A character (data is a vector) or vector of characters (data is a matrix) corresponding to the colours to use. Default set to \code{NULL} in which case \link[grDevices]{rainbow} is used with 50\% transluency.}

\item{border}{A character (data is a vector) or vector of characters (data is a matrix) corresponding to the colours to use. Default set to \code{NULL} in which case \link[grDevices]{rainbow} is used without transluency.}

\item{lwd}{A numeric (data is a vector) or vector of numerical values (data is a matrix) corresponding to the border thickness to use. Default to 1.}

\item{points}{A \link[base]{logical} (with default set to \code{TRUE}) indicating if individual points should be drawn, or a list with names \code{col}, \code{pch}, \code{cex}, respectively indicating i/ the colour(s), ii/ the symbol(s) , iii/ the expension factor(s) of the points, and iv/ the level of jittering for the points on the x-axis. The input \code{points = TRUE} is equivalent to \code{points = list(col=NULL, pch=16, cex=1, jitter=0)}. \code{jitter=0.25} add a uniform noise with distribution \code{U[-0.25,0.25]} to the data (might be useful for counts). When \code{col=NULL} (the default), the same colour as for the non-parametric density plot is used.}

\item{first}{A character indicating if the \code{points} or the non-paramteric density plot should be plotted first. Default to \code{points}.}
}
\description{
Adds non-paramteric density plots to a plot (typically created by \link{.ep}).
}
\examples{
\dontrun{
  data = data.frame(norm=rnorm(500,rep(c(0,5),each=250),1),
                   unif=stats::runif(500,-2,6))    
 .ep(range(data), c(0, 0.75))
  axis(1)
  axis(2, las = 2)
 .hist(data)
} 
}
\seealso{
\link[grDevices]{rainbow}, \link[sm]{sm.density}
}
