% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iboot.r
\name{.iboot.conv.lm}
\alias{.iboot.conv.lm}
\title{IB final convergence assessment}
\usage{
.iboot.conv.lm(estimates, alpha)
}
\arguments{
\item{estimates}{the (k x p) matrix of estimates, where k and p respecively denotes the number of iterations and total number of parameters.}

\item{alpha}{the p-value cut-off value used to assess significance of the slope of each target parameter.}
}
\value{
a logical
}
\description{
Linear-model based assessment of IB convergence. In complex models and/or in models considering discrete responses, the same response vector can sometimes be obtained via different sets of parameters, often leading the implicit boostrap to go from a such a set to the next one between iterations. In such cases, fitting a linear model on parameter estimate vectors leads to slopes close to 0. It is what this function tests, typically when iteration K is reached, i.e, when k=K.
}
