% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iboot.r
\name{.iboot.conv.relative}
\alias{.iboot.conv.relative}
\title{IB convergence assessment}
\usage{
.iboot.conv.relative(estimates, tol)
}
\arguments{
\item{estimates}{the (k x p) matrix of estimates, where K and p respecively denotes the maximum number of iterations and total number of parameters.}

\item{tol}{a convergence relative tolerance value. Default is tol = 1e-3.}
}
\value{
a logical
}
\description{
Assessment of IB convergence by comparing estimates at iteration k to estimates at iteration k-1 in relative terms and checking if the sum of the squared relative difference is smaller than the tolerence value.
}
