% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iboot.r
\name{.iboot}
\alias{.iboot}
\title{IB estimator}
\usage{
.iboot(
  data,
  est = NULL,
  est.control = NULL,
  gen = NULL,
  gen.control = NULL,
  conv = .iboot.conv.relative,
  conv.control = list(tol = 0.001),
  convK = .iboot.conv.lm,
  convK.control = list(alpha = 0.01),
  B = 1999,
  B1 = 0,
  R = 49,
  K = NULL,
  parallel = TRUE,
  n.cores = 9,
  print = TRUE
)
}
\arguments{
\item{data}{a dataset, typically of class data.frame, with observations in rows and relevant information (outcomes, predictors, grouping factors, ...) in columns.}

\item{est}{the (possibly biased initial) estimator (self-defined) function. Its input should take '\code{data}' as first argument and, optionally, be followed by elements indicated under argument '\code{est.control}'. The output should be a named vector of target parameter estimates.}

\item{est.control}{An optional list of parameters for the estimator function indicated under '\code{est}'.}

\item{gen}{a (self-defined) data generating function. Its input should take '\code{seed}' and '\code{par}' as first and second arguments and be followed, optionally, by elements indicated under argument '\code{gen.control}'. The output should be a dataset 'similar' to the one indicated under argument '\code{data}' and could be used as input in the function specified under '\code{est}'. '\code{par}' should be a vector of parameters of same length and names as the output of the function '\code{est}'.}

\item{gen.control}{An optional list of parameters for the estimator function indicated under '\code{gen}'.}

\item{conv}{a (self-defined) function assessing convergence based on the estimates obtained on iterations 1 to k (where k denotes the iteration at which convergence is assessed). Its output should be logical. Its first argument should be the (k x p) matrix of estimates, and could be followed, optionally, by elements indicated under argument '\code{conv.control}'. Check \link{.iboot.conv.relative} (default) and \link{.iboot.conv.absolute} for examples. In complex models, (co)variance parameters are typically more variable than others and may not systematically be estimated in all cases (due to convergence issues, for example), so that a convergence function specific to the problem at hand may be more suitable that the suggested ones.}

\item{conv.control}{An optional list of parameters for the convergence function indicated under '\code{conv}'. Default is a list with argument 'tol' = 1e-3, where 'tol' denotes the tolerance value used in \link{.iboot.conv.relative} (default).}

\item{convK}{a (self-defined) function assessing convergence based on the estimates obtained on iterations 1 to K (where K denotes the maximum number of iterations). Its output should be logical. Its first argument should be the (K x p) matrix of estimates, and could be followed, optionally, by elements indicated under argument '\code{conv.control}'. In complex models and/or in models considering discrete responses, the same response vector can sometimes be obtained via different sets of parameters, often leading the implicit boostrap to go from a such a set to the next one between iterations, thus justifying a different assessment of convergence in such cases. Check \link{.iboot.conv.lm} (default) for an example.}

\item{convK.control}{An optional list of parameters for the convergence function indicated under '\code{convK}'. Default is a list with argument 'tol' = 1e-3, where 'tol' denotes the tolerance value used in \link{.iboot.conv.relative} (default).}

\item{B}{either a scalar indicating the number of IB samples which would be generated with seed 1:B or a list of seeds to be used to generate the IB samples. Default is B = 1999.}

\item{B1}{the first seed, i.e., seed number 1, where the total number of (potential) seeds equals B x R. IMPORTANT: To avoid seed overlap, B1 is internally multiplied by '10^nchar(BR)'. Default = 0.}

\item{R}{the number of alternative seeds to consider assuming the seed of interest does not lead to estimates (no convergence). Default is R=49 (so that 50 seeds per task are considered in total).}

\item{K}{the maximum number of iterations to seek convergence for a given bootrap sample. If K=NULL (default), K is set to ceiling(10*log(N)) where N denotes the number of rows of the dataset provided in argument 'data'.}

\item{parallel}{a logical indicating if parallel computation should be used. If parallel = TRUE (default), package parallel with forking (ie, not possible for Windows at the moment).}

\item{n.cores}{number of cores for parallel computation. Default to 9.}

\item{print}{a logical indicating if progress should be printed (useful for debugging). Default is print = TRUE.}
}
\value{
The function \link{.iboot} returns an S3 object of class 'iboot' with available print/summary/plot functions.
}
\description{
IB estimator
}
