% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iboot.r
\name{.iboot_seed}
\alias{.iboot_seed}
\title{implicit bootrap sample related estimates}
\usage{
.iboot_seed(
  seed,
  par,
  est = est,
  est.control = est.control,
  gen = gen,
  gen.control = gen.control,
  conv = conv,
  conv.control = conv.control,
  convK = convK,
  convK.control = convK.control,
  n.seed = n.seed,
  R = R,
  K = K,
  plot = FALSE,
  print = print
)
}
\arguments{
\item{seed}{the (first!) seed related to the sample of interest. If data generation and estimation related to this seed does not lead to convergence, 50 other seeds are used. This list of seeds is defined as seq(seed,seed+n.seed*R,n.seed) where n.seed is defined further below.}

\item{par}{the vector of parameters as estimated by 'est' on the original dataset.}

\item{est}{refer to \link{.iboot} for details.}

\item{est.control}{refer to \link{.iboot} for details.}

\item{gen}{refer to \link{.iboot} for details.}

\item{gen.control}{refer to \link{.iboot} for details.}

\item{conv}{refer to \link{.iboot} for details.}

\item{conv.control}{refer to \link{.iboot} for details.}

\item{convK}{refer to \link{.iboot} for details.}

\item{convK.control}{refer to \link{.iboot} for details.}

\item{n.seed}{number of IB samples.}

\item{R}{refer to \link{.iboot} for details.}

\item{K}{refer to \link{.iboot} for details.}

\item{plot}{only for check. Default to FALSE.}

\item{print}{refer to \link{.iboot} for details.}
}
\value{
The function \link{.iboot_seed} returns the vector of implicit bootrap sample related estimates.
}
\description{
internal function
}
