% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.r
\name{.idc}
\alias{.idc}
\title{combines id objects}
\usage{
.idc(name = "comb", ..., sep = "|", fill = TRUE, ask = TRUE)
}
\arguments{
\item{name}{name (by default 'comb' for 'combination')}

\item{...}{list of id objects}

\item{sep}{the separator to be used for the id column of the resulting object}

\item{fill}{either a (single) character or a logical. If not FALSE, all elements of the id column have the same length and are filled with either a blank space (TRUE) or the character of choice.}

\item{ask}{a logical. When ask equals TRUE (default), the function checks if an r object of .GlobalEnv is already named 'id.X' or 'n.X' (where X is the name specified in the previous argument) exists and, if it does, ask for the authorisation to overwrite them. The question is not asked when options()$dotfunctions_dontask == TRUE, where options()$dotfunctions_dontask can be set to TRUE (or FALSE) by running options(dotfunctions_dontask = TRUE) in .First, for example, and is set to TRUE in an R session if the user once answers 'always' to the question.}
}
\value{
a data frame with columns 'pos' (scalar vector going from 1 to the number of rows), 'id' (character vector ordered by numerical value whe the input is numerical) and as many columns as there were (unique) id objects to combine, each of them being identifiable via the id extension name (i.e, "x" for id.x, for example) . the returned object is of class 'data.frame' and '.idf'
}
\description{
creates an id object based on all combinations of levels of id objects of interest (created by means of .idf)
}
\examples{
   
\dontrun{
.idf(table(rpois(10,3)),"X")
.idf(c("a","b"),"Y")
.idc("cYX",id.Y,id.X,sep="&")
print(id.cYX)
print(n.cYX)
}
}
