% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.r
\name{.ide}
\alias{.ide}
\title{list of existing id elements}
\usage{
.ide()
}
\value{
output is a string vector.
}
\description{
identify all existing id objects (typically created by means of .idf in an environment (typically .GlobalEnv).
}
\examples{
   
\dontrun{    
.ide()
}    
}
