% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{.isna}
\alias{.isna}
\title{plot missing pattern}
\usage{
.isna(
  data,
  col = c("light gray", "red"),
  transpose = FALSE,
  order.row = TRUE,
  order.col = TRUE,
  cex.axis = c(1, 0.4),
  hadj = 2.5
)
}
\arguments{
\item{data}{a dataset with missings}

\item{col}{a vector of colours, the first one for the non-missing information, the second one for the missing one}

\item{transpose}{a logical indicating if the dataset should be transposed.}

\item{order.row}{a logical indicating if the rows of the dataset should be ordered according to a distance functions,}

\item{order.col}{a logical indicating if the columns of the dataset should be ordered according to a distance functions,}

\item{cex.axis}{a numerical vector of length 2 indicating the font size on the ast and second axes,}

\item{hadj}{a numerical scalar indicating the hadj for the sample size onformation on axis 2,}
}
\description{
plot missing pattern
}
\examples{
\dontrun{
x   = matrix(sample(c(NA,1),100,replace=TRUE),ncol=4,dimnames=list(1:25,LETTERS[1:4]))
.isna(x)
}
}
