% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.r
\name{.list}
\alias{.list}
\title{list based on ide file}
\usage{
.list(id, value = NA)
}
\arguments{
\item{id}{a file of class .idf}

\item{value}{content of the list, NA by default}
}
\value{
a list
}
\description{
creates a list with length and names coming from an id file, typically generated via .idf().
}
\examples{
    
\dontrun{       
.idf(LETTERS[1:6],"Y")   
.l(id.Y)    
}   
}
