% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme.r
\name{.lmecov}
\alias{.lmecov}
\title{lme fit random effect covariance matrix}
\usage{
.lmecov(fit, residual.variance = TRUE)
}
\arguments{
\item{fit}{a fit of class "lme"}

\item{residual.variance}{a logical indicating if the residual variance should be included or not. By default, residual.variance=TRUE so that the residual variance is included and the ouput is of size (q x q), where q denotes the number of random effects plus 1 (for the residual variance).}
}
\value{
the covariance matrix (of class matrix)
}
\description{
extract the estimated random effect covariance matrix of a linear mixed model fitted by means of nlme::lme().
}
\examples{
\dontrun{
# library(nlme)
fm1 <- lme(distance ~ age, data = Orthodont) # random is ~ age
fm2 <- lme(distance ~ age + Sex, data = Orthodont, random = ~ 1)
.lmecov(fm1)
.lmecov(fm2)
}
}
