% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.logit}
\alias{.logit}
\title{shortcut}
\usage{
.logit(p)
}
\arguments{
\item{p}{vector of rational numbers belonging to (0,1)}
}
\value{
a vector of rational numbers
}
\description{
defines the .logit function on a vector of numerical values belonging to (0,1). typically used with probabilities related to a linear regression.
}
\examples{
.logit(0.5)
}
