% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.nf}
\alias{.nf}
\title{.nf}
\usage{
.nf(file)
}
\arguments{
\item{file}{a dataframe}
}
\description{
.nf() stands for 'No (bloody) factors' and change columns of class 'factor' of a data frame into columns of class 'character'.
Since read.csv() \link{and equivalent functions} changes the default of the argument 'stringsAsFactors' to FALSE, .nf() is less useful.
Returns a data frame of same size as input but without factors.
}
