% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.p}
\alias{.p}
\title{shortcut}
\usage{
.p(..., sep = "")
}
\arguments{
\item{...}{objects}

\item{sep}{a character string to separate the terms}
}
\description{
lazy shortcut for paste0
}
\examples{
.p(LETTERS[1:10],collapse="+")
}
