% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{.polygon}
\alias{.polygon}
\title{polygon 'shaded' area}
\usage{
.polygon(x, y1, y2 = 0, col = gray(0.75), border = NA, ...)
}
\arguments{
\item{x}{vector of the x axis}

\item{y1}{vector of values of f(x) for the first bound (i.e., upper or lower). It should be a vector of the same length as x.}

\item{y2}{value(s) of f(x) for the second bound. It can either be a vector of the same length as y1 or a scalar of class numeric (default y2 = 0).}

\item{col}{the colour to be used for the surface}

\item{border}{the colour to be used for the border (default border=NA, i.e., no border)}

\item{...}{other parameter sent to \link[graphics]{polygon}}
}
\description{
add a coloured area defined by the function polygon to an existing plot. Ideal to display confidence bounds.
}
\examples{
\dontrun{
x   = seq(-4,4,length=1000)
f.x = dnorm(x)
.ep(xlim=range(x),ylim=range(f.x))
axis(1, pos=0)
axis(2)
lines(x,f.x)  
posw = x>-1
.polygon(x[posw],y1=f.x[posw])
}
}
