% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicklook.r
\name{.ql}
\alias{.ql}
\title{.ql}
\usage{
.ql(
  obj,
  dim1 = 10,
  dim2 = 6,
  dim3 = 2,
  length.vector = 24,
  length.list = 5,
  digit = 3,
  inf.col = "all",
  inf.list = "sample",
  full = TRUE
)
}
\arguments{
\item{obj}{R object}

\item{dim1}{max number of rows}

\item{dim2}{max number of columns}

\item{dim3}{max number of deepth (array)}

\item{length.vector}{max length of a vector}

\item{length.list}{max length of a list}

\item{digit}{rounding}

\item{inf.col}{"all", "sample","no"}

\item{inf.list}{"all", "sample","no"}

\item{full}{by default is TRUE}
}
\description{
allows a better look on object
}
