% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateandtime.r
\name{.sectotime}
\alias{.sectotime}
\title{time function}
\usage{
.sectotime(n.sec)
}
\arguments{
\item{n.sec}{vector of integers (ranging from 0 to 86400, the number of seconds in a day).}
}
\value{
output is a string vector of format 'hh:mm:ss'.
}
\description{
defines the time (hh:mm:ss) corresponding to each element of a vector of seconds
}
\examples{
.sectotime(c(1000,10000))
}
