% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateandtime.r
\name{.timetosec}
\alias{.timetosec}
\title{second function}
\usage{
.timetosec(x)
}
\arguments{
\item{x}{string vector of integers 'hhmmss'.}
}
\value{
output is a vector of integers.
}
\description{
defines the second (integer ranging from 0 to 86400, the number of seconds in a day) corresponding to a vector of integers of format 'hhmmss' (where 024640 or 24640 correspond to '02:46:40')
}
\examples{
.timetosec(c(001640,24640))
}
