% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.trunc}
\alias{.trunc}
\title{shortcut}
\usage{
.trunc(x)
}
\arguments{
\item{x}{input vector of strings}
}
\value{
a vector of strings
}
\description{
removes the empty spaces coming after the last character for each element of a vector of strings and thus transforms, for example, "abc def gh     " into "abc def gh".
}
\examples{
.trunc("abc def gh     ")
}
