% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{.violin}
\alias{.violin}
\title{Violin plot}
\usage{
.violin(
  data,
  at = NULL,
  h = 1,
  violin = TRUE,
  boxplot = TRUE,
  horizontal = FALSE,
  points = TRUE,
  first = "points"
)
}
\arguments{
\item{data}{A vector or matrix of numerical values.}

\item{at}{A scalar if \code{data} is a vector, or a vector of scalars if \code{data} if a matrix. When \code{at=NULL} (default), value \code{1} and \code{1:p} are respectively used if data is a vector or a matrix, where \code{p} denote the number of columns of the matrix.}

\item{h}{A numeric corresponding to the smoothing parameter. Default to 1. When \code{h=NULL}, the smoothing value is optimised in the \link[sm]{sm.density} function (Good luck with that).}

\item{violin}{A \link[base]{logical} (with default set to \code{TRUE}) indicating if the violin plot should be drawn, or a list with names \code{col}, \code{border}, and \code{lwd}, respectively indicating i/ the background colour, ii/ the colour of the border, and iii/ the tickness of the border. The input \code{violin = TRUE} is equivalent to \code{violin = list(col="#8B897040", border=gray(0),lwd=1)}.}

\item{boxplot}{A \link[base]{logical} (with default set to \code{TRUE}) indicating if a boxplot should be drawn, or a list with names \code{col}, \code{border}, \code{width} and \code{lwd}, respectively indicating i/ the colour of the boxplot, ii/ the colour of the border of the boxplot, iii/ the boxplot width, and iv/ the tickness of the boxplot border. The input \code{boxplot = TRUE} is equivalent to \code{boxplot = list(col="#8B897040", border=gray(.2),width =.05,lwd=1)}.}

\item{horizontal}{A \link[base]{logical} (with default set to \code{FALSE}) indicating if the plot should be horizontal or vertical.}

\item{points}{A \link[base]{logical} (with default set to \code{TRUE}) indicating if individual points should be drawn, or a list with names \code{col}, \code{pch}, \code{cex}, respectively indicating i/ the colour(s), ii/ the symbol(s) , iii/ the expension factor(s) of the points, and iv/ the level of jittering for the points on the y-axis. The input \code{points = TRUE} is equivalent to \code{points = list(col="#2E008B", pch=16, cex=1, jitter=0)}. \code{jitter=0.25} add a uniform noise with distribution \code{U[-0.25,0.25]} to the data (might be useful for counts).}

\item{first}{A character indicating if the \code{points} or the violin plot should be plotted first. Default to \code{points}.}
}
\description{
Adds violin density estimates to a plot (typically created by \link{.ep}).
}
\examples{
\dontrun{
  data = data.frame(norm=rnorm(500,rep(c(0,5),each=250),1),
                   unif=stats::runif(500,-2,6))    
 .ep(c(.5,2.5), range(data))
  axis(1, at  = 1:ncol(data), colnames(data))
  axis(2, las = 2)
 .violin(data)
} 
}
\seealso{
\link[sm]{sm.density}
}
