% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.r
\name{.w}
\alias{.w}
\title{warning function}
\usage{
.w(x = NULL)
}
\arguments{
\item{x}{warning message (optional)}
}
\value{
warning message (character)
}
\description{
prints a visually striking error message
}
\examples{
.w()
}
