% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls.r
\name{.xls}
\alias{.xls}
\title{.xls}
\usage{
.xls(file, sheet = NULL, skip = 0, col_names = TRUE, col_types = "guess", ...)
}
\arguments{
\item{file}{input Excel file}

\item{sheet}{sheet}

\item{skip}{skip rows}

\item{col_names}{column names}

\item{col_types}{columnt types}

\item{...}{...}
}
\description{
Import xls data
}
