% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iboot.r
\name{print.iboot}
\alias{print.iboot}
\title{print function}
\usage{
\method{print}{iboot}(x, ...)
}
\arguments{
\item{x}{object of class 'iboot'.}

\item{...}{additional arguments to be passed nowhere.}
}
\value{
print.
}
\description{
print function for objects of class 'iboot'
}
