% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iboot.r
\name{print.summary.iboot}
\alias{print.summary.iboot}
\title{print function}
\usage{
\method{print}{summary.iboot}(x, ...)
}
\arguments{
\item{x}{object of class 'summary.iboot'.}

\item{...}{additional arguments to be passed to print.}
}
\value{
print.
}
\description{
print function for objects of class 'summary.iboot'
}
