% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iboot.r
\name{summary.iboot}
\alias{summary.iboot}
\title{summary function}
\usage{
\method{summary}{iboot}(
  object,
  point = median,
  point.control = list(na.rm = FALSE),
  interval = quantile,
  interval.control = list(prob = c(0.025, 0.975), na.rm = FALSE),
  ...
)
}
\arguments{
\item{object}{object of class 'iboot'}

\item{point}{a function defining the point estimate for a target parameter. Its first argument must be the vector IB estimates of length B and could be followed, optionally, by elements indicated under argument '\code{point.control}'. Default is the median function.}

\item{point.control}{An optional list of parameters for the point estimate function indicated under '\code{point}'. Default is \code{na.rm = FALSE}.}

\item{interval}{a function defining the interval estimates for a target parameter. Its first argument must be the vector IB estimates of length B and could be followed, optionally, by elements indicated under argument '\code{point.control}'. The output should be a named vector (typically of length 2 but longer if wished) of estimates. Default is the quantile function leading to percentile confidence intervals. Other choices may lead to better coverages when B is small to moderate.}

\item{interval.control}{An optional list of parameters for the interval estimate function indicated under '\code{interval}'. Defaults are \code{probs = c(0.025,.975)}, leading to confidence intervals at the 0.95 levels, and \code{na.rm = FALSE}.}

\item{...}{additional arguments to be passed nowhere.}
}
\value{
an object of class 'iboot.summary' with available print function. The object is a list with elements
\itemize{
\item initial - The vector of initial target parameter estimates.
\item boot - The (B x p) matrix of IB estimates.
\item final - The vector of final target parameter estimates (as estimated by the function \code{point} in summary).
\item ci - The (p x 2) matrix of confidence intervals)..
}
}
\description{
summary function for objects of class 'iboot'
}
