% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{violin.density}
\alias{violin.density}
\alias{.violin.density}
\title{Violin plot density estimates}
\usage{
.violin.density(data, h = NULL)
}
\arguments{
\item{data}{data vector.}

\item{h}{a scalar corresponding to the smoothing parameter.}
}
\value{
a list of coordinates and corresponding density values.
}
\description{
Extracts density estimates for a violin plot.
}
\seealso{
\link[sm]{sm.density}
}
