% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.adf}
\alias{.adf}
\title{shortcut}
\usage{
.adf(matrix, colnames = NULL, rownames = NULL)
}
\arguments{
\item{matrix}{input matrix}

\item{colnames}{colnames}

\item{rownames}{rownames}
}
\value{
a data frame.
}
\description{
change class of a matrix to 'data.frame'. Less busy people may consider the function as.data.frame().
}
\examples{
.adf(matrix(1:10,ncol=2),colnames=c("a","b")) 
}
