% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{.axis}
\alias{.axis}
\title{axis function}
\usage{
.axis(
  data,
  split,
  side = 3,
  at.y = NULL,
  col.abline = "gray",
  col.label = "blue",
  col.axis = "blue",
  padj = NA,
  hadj = NA,
  cex = 1,
  add.abline = TRUE,
  add.label = TRUE,
  add.axis = FALSE,
  label = NULL,
  ...
)
}
\arguments{
\item{data}{a file ('id.x'-like) with a column 'pos' (position on the xaxis), typically 1 to nrow for a split at the full file level}

\item{split}{a factor (typically year/month, school/class, aso)}

\item{side}{3 = over, 1 = below}

\item{at.y}{if non-null, label are displayed in the plot on an horizontal line
at 'at.y' on the y-axis}

\item{col.abline}{colour of the vertical boundaries in the argument 'col' of \link[graphics]{abline}}

\item{col.label}{colour of the label}

\item{col.axis}{colour of the x-axis if add.axis == TRUE}

\item{padj}{padj}

\item{hadj}{hadj}

\item{cex}{cex}

\item{add.abline}{TRUE for vertical boundaries}

\item{add.label}{TRUE for label}

\item{add.axis}{TRUE for x-axis}

\item{label}{label}

\item{...}{other useful arguments: cex for the cex of labels in axis() and text(), lty for the lty in abline(), padj or hadj for the adjustment of the location of the label in axis()}
}
\description{
create fancy plot axes (typically based on an id file).
}
