% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.cat}
\alias{.cat}
\title{for loop iteration print}
\usage{
.cat(i, n.loop, n.dotline = 50, msg = NULL)
}
\arguments{
\item{i}{iteration number}

\item{n.loop}{number of loop}

\item{n.dotline}{number dot per line}

\item{msg}{message text}
}
\description{
cat a dot or an iteration number at each iteration of a for loop. Idea is to see progress.
}
\examples{
for(i in 1:200){.cat(i,200,10)}
}
