% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediapulse.r
\name{.cbf}
\alias{.cbf}
\title{cbind files function}
\usage{
.cbf(list.files, path = F, name.loaded.file = "mx.g.zt")
}
\arguments{
\item{list.files}{list of files to cbind}

\item{path}{indicate path=TTRUE to load a file and path=FALSE if files are in GlobalEnv}

\item{name.loaded.file}{name of the loaded files (typically 'mx.x.zt', for example)}
}
\value{
combined matrix or data frame.
}
\description{
function used in radio and tv-research to combine files
}
