% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{.circle}
\alias{.circle}
\title{circle function}
\usage{
.circle(x, plot = TRUE, ...)
}
\arguments{
\item{x}{x is a vector defined as follows: first element =  centre x-coordinate, second element =  centre y-coordinate, third element = radius.}

\item{plot}{a logical. use TRUE to plot or FALSE to save dataset of x and y coordinates of 200 points}

\item{...}{...}
}
\description{
plot a circle based on coordinates of centre and radius. Note that circles may look like ellipses if the aspect ratio (argument 'asp' of plot) is different from 1
}
\examples{
\dontrun{
.ep(asp=1)
.circle(c(.5,.5,.25))
}
}
