% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{.ep}
\alias{.ep}
\title{empty plot function}
\usage{
.ep(xlim = c(0, 1), ylim = c(0, 1), ...)
}
\arguments{
\item{xlim}{xlim}

\item{ylim}{ylim}

\item{...}{...}
}
\description{
generate an empty plot (no axes, no axe labels, no box, no points)
}
\examples{
\dontrun{
.ep()
}
}
