% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.fill}
\alias{.fill}
\title{shortcut}
\usage{
.fill(x, npos = NULL, with = " ", front = TRUE)
}
\arguments{
\item{x}{input vector of strings}

\item{npos}{number of position to add, default to the max of the input vector}

\item{with}{character to add (default is an empty space)}

\item{front}{a logical indicating if the chracter is to add at the front (default)}
}
\value{
a vector of strings (with added characters)
}
\description{
for printing purposes, add characters of choice (0s or empty spaces, for example) to a vector of string. typically used to ensure ids have the same number of characters
}
\examples{
.fill(1:20,with="0")
}
