% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installpackages.r
\name{.ip}
\alias{.ip}
\title{.ip}
\usage{
.ip(pkg, mirror = 1, dependencies = TRUE, quiet = FALSE)
}
\arguments{
\item{pkg}{name of a package}

\item{mirror}{source mirror}

\item{dependencies}{dependencies}

\item{quiet}{suppress messages}
}
\description{
note: files are stored under ">.libPaths"
}
