% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.nf}
\alias{.nf}
\title{.nf}
\usage{
.nf(file)
}
\arguments{
\item{file}{a dataframe or a matrix}
}
\description{
the function ".nf" for "No Factor" transform "factor-vector
of dataframe in "character" (when importing a dataset with read,
more elegant to use stringsAsFactors=FALSE)
}
