% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divers.r
\name{.nf}
\alias{.nf}
\title{factor exterminator}
\usage{
.nf(file)
}
\arguments{
\item{file}{a dataframe}
}
\value{
returns a data frame of same size as input but without factors.
}
\description{
change columns of class 'factor' of a data frame into columns of class 'character'. 'nf' stands for 'No (bloody) factors'. Since read.csv() (and equivalent functions) changes the default of the argument 'stringsAsFactors' to FALSE, .nf() is less useful.
}
\examples{
sapply(.nf(data.frame(y=rnorm(10),x=factor(rep(0:1,each=5)))),class)
}
