% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable.r
\name{.pval}
\alias{.pval}
\title{.pval}
\usage{
.pval(pval, digit = 4)
}
\arguments{
\item{pval}{p-value}

\item{digit}{number of digits for rounding rule}
}
\description{
For printing purposes, change 1e-299 into <0.0001
}
