% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.r
\name{.pval}
\alias{.pval}
\title{p-val function}
\usage{
.pval(pval, digit = 4)
}
\arguments{
\item{pval}{vector of p-values}

\item{digit}{number of digits for rounding rule}
}
\value{
output is a vector of class 'character' with a fixed number of postions after the comma.
}
\description{
print function for p-values, change 1e-299 into <0.0001.
}
\examples{
.pval(c(0,0.01,0.05,0.1,1))    
}
