% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicklook.r
\name{.ql}
\alias{.ql}
\title{quick look function}
\usage{
.ql(
  obj,
  dim1 = 10,
  dim2 = 6,
  dim3 = 3,
  length.vector = 24,
  length.list = 5,
  digit = 3,
  inf.col = "all",
  inf.list = "sample",
  full = TRUE
)
}
\arguments{
\item{obj}{R object}

\item{dim1}{max number of rows to display}

\item{dim2}{max number of columns to display}

\item{dim3}{max number of dimensions to display for array}

\item{length.vector}{max length of a vector}

\item{length.list}{max length of a list}

\item{digit}{rounding}

\item{inf.col}{"all" (default), "sample","none"}

\item{inf.list}{"all" (default), "sample","none"}

\item{full}{by default is TRUE}
}
\description{
generates a 'quick look' of R objects of different classes. Currently handles matrices, data frames, lists and vectors. Typically prints a subset of the object (first rows and columns for matrices, for example) and some relevant summaries (like the class, number of 0s and NAs, for each column of a data frame, for example)
}
\examples{
.ql(.adfm(rnorm(100),ncol=4))
}
