% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.r
\name{.sig}
\alias{.sig}
\title{sig function}
\usage{
.sig(pval)
}
\arguments{
\item{pval}{vector of p-values}
}
\value{
output is a vector of class 'character'.
}
\description{
star system for p-values.
}
\examples{
.sig(c(0,0.01,0.05,0.1,1))
}
