% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.r
\name{.tex}
\alias{.tex}
\title{tex function}
\usage{
.tex(
  input,
  file = NULL,
  cmidrule = NULL,
  col.row = NULL,
  col.col = NULL,
  pos.textbfrow = NULL,
  pos.textbfcol = NULL,
  pos.centeredrow = NULL,
  pos.centeredcol = NULL
)
}
\arguments{
\item{input}{the table, including rownames and colnames}

\item{file}{the name and path of the tex file}

\item{cmidrule}{a matrix of same size as input with non-NA entries
for position with a cmidrule. Example: c(NA,1,1,2,2) for a given row}

\item{col.row}{colour of each row}

\item{col.col}{colour of each column}

\item{pos.textbfrow}{position of the rows with bold names}

\item{pos.textbfcol}{position of the columns with bold names}

\item{pos.centeredrow}{position of the columns which should be centered in row}

\item{pos.centeredcol}{position of the rows which should be centered in column}
}
\description{
defines the centre part of the tex code corresponding to a Table (and typically called, in TeX, via the input function). the input is a matrix with row and column names.
}
