% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.r
\name{.w}
\alias{.w}
\title{warning function}
\usage{
.w(x = NULL, immediate. = FALSE)
}
\arguments{
\item{x}{warning message (optional)}

\item{immediate.}{logical, indicating if the warning should be output immediately. check ?warning for further information on this argument}
}
\value{
warning message (character)
}
\description{
prints a visually striking error message
}
\examples{
\dontrun{
.w()
}
}
